Extranonce 2 variable length (#1168)
* ignore debug.log

* fix extranonce_2_generate to handle variable length extranonce_2 for #1167

* added extranonce_2_length clamping to 32 bytes

---------

+++ components/stratum/include/mining.h
@@ -36,7 +36,7 @@ bm_job construct_bm_job(mining_notify *params, const char *merkle_root, const ui
 
 double test_nonce_value(const bm_job *job, const uint32_t nonce, const uint32_t rolled_version);
 
-char *extranonce_2_generate(uint32_t extranonce_2, uint32_t length);
+char *extranonce_2_generate(uint64_t extranonce_2, uint32_t length);
 
 uint32_t increment_bitmask(const uint32_t value, const uint32_t mask);
